python_service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
torch==2.5.1
torchvision==0.20.1
opencv-python==4.10.0.84
pillow==11.0.0
numpy==2.1.3
python-multipart==0.0.18
